backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-cloud-pubsub==2.21.1
pytest==8.2.0
pytest-asyncio==0.23.6
